add: basic x11 support

+++ minigox.c
@@ -9,7 +9,7 @@
 static struct CharInfo curr;
 static struct Method method;
 static struct {
-    long x, y;
+    double x, y;
 } mouse;
 static struct {
     int ctrl, alt, shift, caps, num, scroll;
@@ -46,6 +46,8 @@ static bool minigox_process_char(char key) {
 
 #if defined(MINIGOX_WIN32)
 #include "minigox_win32.h"
+#elif defined(MINIGOX_X11)
+#include "minigox_x11.h"
 #else
 #error "No implementation found"
 #endif
